// src/features/conversion/domain/enums/ConversionType.ts
export enum ConversionType {
  // Shortcuts
  ID_CARD = 'id-card',
  SCAN_DOCS = 'scan-docs',
  BOOK_SCAN = 'book-scan',
  QR_CODE = 'qr-code',
  TO_TEXT = 'to-text',

  // Process Files
  IMPORT_FILES = 'import-files',
  COLLAGE_IMAGES = 'collage-images',
  WATERMARK_PDF = 'watermark-pdf',
  MERGE_PDF = 'merge-pdf',
  COMPRESS_PDF = 'compress-pdf',

  // PDF Tools
  PDF_SIGNATURE = 'pdf-signature',
  PDF_PASSWORD = 'pdf-password',
  PAGE_REORDER = 'page-reorder',
  PDF_TO_WORD = 'pdf-to-word',
  WORD_TO_PDF = 'word-to-pdf',
  SPLIT_PDF = 'split-pdf',
  IMAGES_TO_PDF = 'images-to-pdf',
}